fix: update justfile paths and refactor imports in Justchat module
- Corrected the path for the `build-move-famcoin` command in the justfile.
- Refactored import statements in `justchat.ts` and `justchat.test.ts` to use the new module structure.
- Removed deprecated `champMarket.ts` and `mockcoins.ts` files to clean up the codebase.
- Updated test files to utilize dry run transactions for better testing practices.
