+++ sdk/test/mockcoins.test.ts
@@ -23,13 +23,9 @@ describe('Mock Coins Tests', () => {
     tx.transferObjects([redCoin], alice.toSuiAddress());
 
     const txBytes = await tx.build({ client: suiClient });
-    const signature = await alice.signTransaction(txBytes);
-    const result = await suiClient.executeTransactionBlock({
+    const dryRunResult = await suiClient.dryRunTransactionBlock({
       transactionBlock: txBytes,
-      signature: signature.signature,
-      options: { showEffects: true, showObjectChanges: true },
     });
-
-    expect(result.effects?.status.status).toBe('success');
+    expect(dryRunResult.effects.status.status).toBe('success');
   });
 });
